refactor(shared): Split enemies.ts into faction-based modules for LLM optimization
Reorganize enemy definitions from a single 648-line file into category-based
modules organized by enemy faction/token color:

- enemies/types.ts: EnemyResistances, EnemyDefinition, EnemyColor interfaces
- enemies/abilities.ts: ABILITY_* constants and ABILITY_DESCRIPTIONS
- enemies/resistances.ts: NO_RESISTANCES, PHYSICAL_RESISTANCE, etc.
- enemies/green.ts: Marauding Orcs (Diggers, Prowlers, etc.)
- enemies/gray.ts: Keep garrison (Crossbowmen, Guardsmen, etc.)
- enemies/brown.ts: Dungeon monsters (Minotaur, Gargoyle, etc.)
- enemies/violet.ts: Mage Tower (Monks, Illusionists, Mages)
- enemies/white.ts: City garrison (Altem Guardsmen, Altem Mages, etc.)
- enemies/red.ts: Draconum dragons
- enemies/index.ts: Re-exports + ENEMIES aggregate + getEnemy()

This follows the same pattern as events.ts and resolveEffect.ts refactoring.
Each module includes documentation with enemy descriptions and role.
