+++ packages/shared/src/actions.ts
@@ -6,7 +6,7 @@ import type { HexCoord, HexDirection } from "./hex.js";
 import type { CardId, SkillId, BasicManaColor, ManaColor } from "./ids.js";
 import type { UnitId } from "./units.js";
 import type { ManaSourceType, SparingPowerChoice } from "./valueConstants.js";
-import type { EnemyId } from "./enemies.js";
+import type { EnemyId } from "./enemies/index.js";
 import type { CombatType } from "./combatTypes.js";
 import type { Element } from "./elements.js";
 import type { TacticId } from "./tactics.js";
